Animation instances may now have their frames set after the animation has been created.  Previously the frames could only be set in the constructor.  This was modified due to the order in which Initialize and LoadContent methods are called in the game.  I expected the order to be Game.Initialize, GameComponent.Initialize, Game.LoadContent, GameComponent.LoadContent.  However it executes as Game.Initialize, GameComponent.Initialize, GameComponent.LoadContent, Game.LoadComponent.
BaseGame is now generic so that the extending game may be referenced directly.  GameState, GameStateManager, TitleScreen, and PlayScreen were modified accordingly to accept the generic BaseGame.

A new Animations class was added to keep track of all animations that will be used by the SpaceInvaders game.

The sprite sheet that will be used with the game was changed to a much better looking set of sprites than I was able to create...

The SmallInvader class was removed in favor of using a single Invader class from which all invader types may derive.